cloud functions/generate_thumbnail/requirements.txt
google-cloud-storage>=2.8.0
google-cloud-bigquery>=3.10.0
pillow>=9.5.0